feat:: updated the contract with implementations and moved to contract.rs
